test of vaciar ranking works

+++ app/src/main/java/com/mainpakage/tetrix/GameOver.java
@@ -44,26 +44,40 @@ public class GameOver extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_game_over);
         bAux = getIntent().getExtras();
-        adaptedArray = new ArrayList<>();
-        scoreText = (TextView) findViewById(R.id.Score);
-        scoreText.setText("Score: " + bAux.getString("Score"));
-        playerName = (TextView) findViewById(R.id.playerName);
-        picCam = (ImageView) findViewById(R.id.picCam);
-        listRanking = (ListView) findViewById(R.id.listRanking);
-        gameOverText = (TextView) findViewById(R.id.gameOver);
-        rankingText = (TextView) findViewById(R.id.rankingText);
-        butCl = (Button) findViewById(R.id.setName);
-        restablecer = (Button) findViewById(R.id.butRestablecer);
-
         gameMode = bAux.getInt("GameMode");
-        if (gameMode == 0)
-            preferences = getSharedPreferences("RankingDefinitive", Context.MODE_PRIVATE);
-        else
-            preferences = getSharedPreferences("RankingPowerDefinitive", Context.MODE_PRIVATE);
+        adaptedArray = new ArrayList<>();
 
+            scoreText = (TextView) findViewById(R.id.Score);
+            scoreText.setText("Score: " + bAux.getString("Score"));
+            playerName = (TextView) findViewById(R.id.playerName);
+            picCam = (ImageView) findViewById(R.id.picCam);
+            listRanking = (ListView) findViewById(R.id.listRanking);
+            gameOverText = (TextView) findViewById(R.id.gameOver);
+            rankingText = (TextView) findViewById(R.id.rankingText);
+            butCl = (Button) findViewById(R.id.setName);
+            restablecer = (Button) findViewById(R.id.butRestablecer);
+
+            System.out.println("HEEEEEELLLLLOOOOOOOOOOOOOOOOOOOOO");
+            /*rankingText = (TextView) findViewById(R.id.rankingText);
+            listRanking = (ListView) findViewById(R.id.listRanking);
+            restablecer = (Button) findViewById(R.id.butRestablecer);
+        */
+        if(gameMode != -2) {
+
+            if (gameMode == 0)
+                preferences = getSharedPreferences("RankingDefinitive", Context.MODE_PRIVATE);
+            else
+                preferences = getSharedPreferences("RankingPowerDefinitive", Context.MODE_PRIVATE);
+
+            scores = preferences.getStringSet("Scores", null);
+            editor = preferences.edit();
+        }else{
+            scores = new HashSet<>();
+            scores.add("NOMBRE/128");
+        }
         updateRanking();
 
-        if (gameMode != -1) {
+        if (gameMode > -1) {
             dispatchTakePictureIntent();
             pn = (EditText) findViewById(R.id.playerName);
         }else{
@@ -125,32 +139,40 @@ public class GameOver extends AppCompatActivity {
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
         if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
             Bundle extras = data.getExtras();
             Bitmap imageBitmap = (Bitmap) extras.get("data");
-            ImageView picCam=(ImageView)findViewById(R.id.picCam);
+            ImageView picCam = (ImageView) findViewById(R.id.picCam);
             picCam.setImageBitmap(imageBitmap);
         }
     }
 
-    private void viewRanking(){
+    public void viewRanking(){
 
-        butCl.setVisibility(View.GONE);
-        gameOverText.setVisibility(View.GONE);
-        playerName.setVisibility(View.GONE);
-        picCam.setVisibility(View.GONE);
-        scoreText.setVisibility(View.GONE);
-        rankingText.setVisibility(View.VISIBLE);
-        listRanking.setVisibility(View.VISIBLE);
-        restablecer.setVisibility(View.VISIBLE);
-
+        if (gameMode != -2) {
+            butCl.setVisibility(View.GONE);
+            gameOverText.setVisibility(View.GONE);
+            playerName.setVisibility(View.GONE);
+            picCam.setVisibility(View.GONE);
+            scoreText.setVisibility(View.GONE);
+            rankingText.setVisibility(View.VISIBLE);
+            listRanking.setVisibility(View.VISIBLE);
+            restablecer.setVisibility(View.VISIBLE);
+        }else{
+            rankingText.setVisibility(View.VISIBLE);
+            listRanking.setVisibility(View.VISIBLE);
+            restablecer.setVisibility(View.VISIBLE);
+        }
         Collections.sort(ranking);
 
         for (PlayerData p : ranking) {
             scores.add(p.toString());
         }
-        editor.putStringSet("Scores", scores);
-        editor.commit();
+        if(gameMode != -2) {
+            editor.putStringSet("Scores", scores);
+            editor.commit();
+        }
         adaptedArray.clear();
         for (int r = 0; (r < ranking.size()) && (r < 10); r++)
             adaptedArray.add(ranking.get(r).toFormatString());
@@ -160,9 +182,7 @@ public class GameOver extends AppCompatActivity {
 
     }
 
-    private void updateRanking(){
-        scores = preferences.getStringSet("Scores", null);
-        editor = preferences.edit();
+    public void updateRanking(){
 
         ranking = new ArrayList<>();
         if (scores == null) {
@@ -175,12 +195,23 @@ public class GameOver extends AppCompatActivity {
                 ranking.add(pd);
             }
         }
+        if ((gameMode == -2) && (scores.isEmpty())) {
+            Toast toast1 = Toast.makeText(getApplicationContext(), "Ranking vacío", Toast.LENGTH_SHORT);
+            toast1.setGravity(Gravity.CENTER,0,0);
+            toast1.show();
+        }
+        System.out.println("Scores"+scores);
+        System.out.println("Ranking"+ranking);
     }
 
     public void vaciarRanking(View v){
 
-        this.editor.clear();
-        this.editor.apply();
+        if(gameMode != -2) {
+            this.editor.clear();
+            this.editor.apply();
+        }else{
+            this.scores.clear();
+        }
         updateRanking();
         viewRanking();
     }
